pyproject.toml
[project]
name = "leaRning"
version = "0.1.0"
authors = [{name = "Arthur Zopellaro", email = "github@zope.me"}]
description = "R Learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "scapy==2.5.0",
  "pysv @ git+https://github.com/arthurazs/pysv.git",
  "pytrade @ git+https://github.com/arthurazs/pytrade.git",
  "ATPIEC @ git+https://github.com/paulopinheirouff/ATPIEC.git"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.1.5",
  "mypy>=1.7.0",
]

[tool.ruff]
select = [
    "F", "E", "W", "C90", "I", "N", "D", "UP", "YTT", "ANN", "ASYNC", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ",
    "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF",
    "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR",
    "PERF", "RUF",
]
ignore = ["D203", "D213", "FA102"]
extend-ignore = ["D", "FIX002", "TD"]
line-length = 120

[tool.mypy]
exclude = ["build", "venv", ".venv"]

[project.scripts]
cli = "src.__main__:main"
